# src/main/java/com/github/jrcodeza/example/model/error/StandardError.java
package com.github.jrcodeza.example.model.error;

public class StandardError {

	private String message;

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}
}